cuebstat: metrics p7 (pow)
